requirements.txt
Django>=4.2
psycopg2-binary>=2.9
djangorestframework
python-telegram-bot>=20.0
requests
python-dotenv
requests
python-telegram-bot
python-dotenv
sqlalchemy
